Struct wiggle_test::HostMemory

pub struct HostMemory { /* fields omitted */ }

Implementations

impl HostMemory

pub fn new() -> Self

pub fn mem_area_strat(align: u32) -> BoxedStrategy<MemArea>

pub fn invert(regions: &MemAreas) -> MemAreas

Takes a sorted list or memareas, and gives a sorted list of memareas covering the parts of memory not covered by the previous

pub fn byte_slice_strat(
size: u32,
align: u32,
exclude: &MemAreas
) -> BoxedStrategy<MemArea>